requirements.txt
beautifulsoup4>=4.13.4,<5
pandas>=2.3.1,<2.4
python-dotenv>=1.0.1,<2
requests>=2.32.4,<3
streamlit>=1.44.1,<2
groq>=0.24.0,<1
plotly>=5.24.1,<6
lxml>=6.0.0,<7
reportlab>=4.2.2,<5
cloudscraper>=1.2.71,<2
